requirements.txt
torch
torchvision
numpy
openai-clip
tqdm
Pillow
transformers